Add info on models
Very unhappy with the current state of this, but I'm a bit out of my depth right now. Need to spend more time with the models and their configurations to provide proper explanations.

Also adds a stub for building the container, since I've now learned that this is the only way to add BMI modules.